PDI-1450: Reorganize packages and folders only when subcommands exist
